requirements.txt
# RPA Leroy Merlin - Dependências Python
# Processamento de dados
pandas>=1.5.0
openpyxl>=3.0.0

# Google Sheets API
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
chardet>=4.0.0 

# Utilitários (já incluídos no Python)
# pathlib - built-in
# datetime - built-in 
